pyproject.toml
[tool.poetry]
name = "wattwise"
version = "0.1.0"
description = "A service for calculating and managing utility bills."
authors = ["Nikita Matsko <nikmd1306@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
tortoise-orm = {extras = ["asyncpg"], version = "^0.25.1"}
aerich = "^0.9.0"
pydantic = "^2.11.5"
apscheduler = "^3.11.0"
jinja2 = "^3.1.6"
weasyprint = "^65.1"
xlsxwriter = "^3.2.3"
aiogram = "^3.20.0.post0"
uvicorn = "^0.34.3"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
ruff = "^0.11.13"
mypy = "^1.16.0"
tomlkit = "^0.13.3"
python-dateutil = "^2.9.0.post0"
pydantic-settings = "^2.9.1"
types-python-dateutil = "^2.9.0.20250516"

[tool.aerich]
tortoise_orm = "app.core.db.TORTOISE_ORM"
location = "app/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
addopts = [
  "--import-mode=importlib",
]

[tool.mypy]
python_version = "3.12"
packages = ["app"]
plugins = [
  "pydantic.mypy"
]

[[tool.mypy.overrides]]
module = [
    "weasyprint",
    "apscheduler.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
